+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## [ Unreleased ]
 
 * add http client 
+* add janus Gateway, Session and Handle interface